\name{Group_HU}
\alias{Group_HU}
\docType{data}
\title{Class of Human Urine spectra.}
\description{Provides the group (1/2/3) to which belongs each signal/spectrum of the Human Urine database.}
\usage{data("Group_HU")}
\format{
  A vector of length  24.
}

\source{Data provided by Dr. Pascal de Tullio and coworkers of the Pharmaceutical Chemistry Laboratory in the Pharmacy Department of the University of Liege (ULg), Belgium.}

\seealso{
See also \code{\link{FidInfo_HU}} for acquisition parameters and \code{\link{FidData_HU}} for the FIDs.
}


\examples{
data(Group_HU)

}
\keyword{datasets}
