\name{PepsNMR-internal}
\alias{Bspline}
\alias{BsplineWarp}
\alias{Interpol}
\alias{PolyWarp}
\alias{PrincipalComponentAnalysis}
\alias{ReadFid}
\alias{SingleWarp}
\alias{beginTreatment}
\alias{binarySearch}
\alias{checkArg}
\alias{endTreatment}
\alias{fftshift1D2D}
\alias{getArg}
\alias{getDirsContainingFid}
\alias{getTitle}
\alias{indexInterval}
\alias{readParams}
\alias{vec2mat}


\title{Internal PepsNMR Functions}
\description{
  Internal PepsNMR Functions
}
\details{
  These are not to be called by the user (or in some cases are just
                                          waiting for proper documentation to be written ...).
}
\keyword{internal}
